requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
pillow>=11.0,<13.0
numpy>=1.26,<3.0
certifi>=2024.0.0
torch>=2.6.0
torchvision>=0.21.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
